common: add pose and velocity frames enums
